+++ lib/puppet/provider/bgp/quagga.rb
@@ -3,6 +3,8 @@ Puppet::Type.type(:bgp).provide :quagga do
 
   commands :vtysh => 'vtysh'
 
+  mk_resource_methods
+
   @resource_map = {
       :import_check => {
           :default => :disabled,