+++ lib/puppet/provider/bgp_network/quagga.rb
@@ -3,6 +3,8 @@ Puppet::Type.type(:bgp_network).provide :quagga do
 
   commands :vtysh => 'vtysh'
 
+  mk_resource_methods
+
   def self.instances
     debug '[instances]'
     found_config = false