+++ lib/puppet/provider/community_list/quagga.rb
@@ -3,6 +3,8 @@ Puppet::Type.type(:community_list).provide :quagga do
 
   commands :vtysh => 'vtysh'
 
+  mk_resource_methods
+
   def initialize(value)
     super(value)
     @property_flush = {}