+++ lib/puppet/provider/ospf_area/quagga.rb
@@ -15,6 +15,8 @@ Puppet::Type.type(:ospf_area).provide :quagga do
 
   commands :vtysh => 'vtysh'
 
+  mk_resource_methods
+
   def initialize(value={})
     super(value)
     @property_flush = {}